WIP: Refactor 'Segment' to its own module.
